edaboweb/models.py
#!/usr/bin/env python
# coding: utf-8
# Copyright © 2015 Wieland Hoffmann
# License: MIT, see LICENSE for details
from flask.ext.sqlalchemy import SQLAlchemy
from sqlalchemy.dialects.postgresql import JSON

db = SQLAlchemy()


class Playlist(db.Model):
    id = db.Column(db.Integer, primary_key=True)
    data = db.Column(JSON)
